[refactor] #23 userTitle-ui refactor
