+++ src/features/user/constants/userDiagnosis.ts
