+++ src/features/user/constants/userSubTextMap.ts
@@ -1,0 +1,8 @@
+export const userSubTextMap = {
+    '회원 가입': '금융 코치 서비스에 오신 것을 환영합니다',
+    '마이 페이지': '금융 코치 서비스에 오신 것을 환영합니다',
+    '회원정보 수정': '회원님의 정보를 최신 상태로 유지해주세요',
+    '금융 성향 진단': '더 정확한 추천을 위해 간단한 설문에 참여해주세요',
+    '워너비 선택': '되고 싶은 금융 유형을 최대 3개까지 선택하세요',
+    '나의 금융 히스토리': '최근 12개월간의 금융 유형 변화를 확인하세요',
+  }
