+++ src/pages/user/UserDiagnosisPage.vue
@@ -7,6 +7,6 @@
   
 <script setup lang="ts">
 import AppLayout from '@/shared/layout/AppLayout.vue'
-import UserTitle from '@/features/user/ui/UserTitle.vue';
+import UserTitle from '@/shared/ui/molecules/UserTitle.vue';
 import UserDiagnosisList from '@/features/user/ui/user-diagnosis/UserDiagnosisList.vue'
 </script>
