+++ src/pages/user/UserHistoryPage.vue
@@ -7,6 +7,6 @@
   
 <script setup lang="ts">
 import AppLayout from '@/shared/layout/AppLayout.vue'
-import UserTitle from '@/features/user/ui/UserTitle.vue';
+import UserTitle from '@/shared/ui/molecules/UserTitle.vue';
 import UserHistroy from '@/features/user/ui/user-history/UserHistroy.vue';
 </script>
