+++ src/pages/user/UserProfileEditPage.vue
@@ -8,7 +8,7 @@
   
 <script setup lang="ts">
 import AppLayout from '@/shared/layout/AppLayout.vue'
-import UserTitle from '@/features/user/ui/UserTitle.vue';
+import UserTitle from '@/shared/ui/molecules/UserTitle.vue';
 import UserProfileCard from '@/features/user/ui/UserProfileCard.vue';
 import UserProfileEditForm from '@/features/user/ui/user-profile-edit/UserProfileEditForm.vue'
 </script>
