+++ src/pages/user/UserSignupPage.vue
@@ -1,12 +1,14 @@
 <template>
     <AppLayout>
-      <UserTitle label="회원 가입"/>
+      <UserTitle :label="'회원 가입'" :subText="userSubTextMap['회원 가입']" />
       <UserSignupForm />
     </AppLayout>
   </template>
   
 <script setup lang="ts">
 import AppLayout from '@/shared/layout/AppLayout.vue'
-import UserTitle from '@/features/user/ui/UserTitle.vue';
 import UserSignupForm from '@/features/user/ui/user-signup/UserSignupForm.vue';
+import UserTitle from '@/shared/ui/molecules/UserTitle.vue';
+import { userSubTextMap } from '@/features/user/constants/userSubTextMap'
+
 </script>
